Moved init_wandb_logger to mttl.logger
